author of recipe was overrided on update
