--- flask-app/requirements.txt
@@ -1,13 +1,0 @@
-Click==7.0
-Flask==1.0.2
-Flask-Login==0.4.1
-Flask-SQLAlchemy==2.3.2
-Flask-WTF==0.14.2
-itsdangerous==1.1.0
-Jinja2==2.10.1
-MarkupSafe==1.1.0
-python-dotenv==0.9.1
-six==1.11.0
-SQLAlchemy==1.3.3
-Werkzeug==0.14.1
-WTForms==2.2.1
